+++ .htaccess
@@ -1,4 +1,4 @@
 RewriteEngine on
 RewriteCond %{REQUEST_FILENAME} !-f
 RewriteCond %{REQUEST_FILENAME} !-d
-RewriteRule ^(.*)$ src/index.php [L]
+RewriteRule ^(.*)$ public/index.php [L]
