+++ app/models/SubTask.php
@@ -1,38 +1,30 @@
 <?php
 
-namespace models;
+namespace app\models;
 
-use Frame\Model\Model;
 use ReflectionException;
+use Frame\Model\Model;
 
-class Task extends Model
+class SubTask extends Model
 {
 //    public int $id;
 //    public string $title;
 //    public string $beschrijving;
 //    public string $status;
-//    public int $verval_datum; // unix timestamp
-//    public string $prioriteit;
 
     // TODO: Timestamps created_at, updated_at
-
 
     public function __construct(array $attributes = [])
     {
         parent::__construct($attributes);
-        $this->table = 'tasks';
-    }
-
-    public function subTasks(): array
-    {
-        return $this->hasMany(SubTask::class);
+        $this->table = 'sub_tasks';
     }
 
     /**
      * @throws ReflectionException
      */
-    public function users(): array
+    public function task(): Task
     {
-        return $this->belongsToMany(User::class);
+        return $this->belongsTo(Task::class);
     }
 }
