+++ docker-compose.yml
@@ -22,7 +22,6 @@ services:
     ports:
       - "3306:3306"
     volumes:
-      - ./db:/docker-entrypoint-initdb.d
       - db-data:/var/lib/mysql
 
 volumes:
