+++ src/Frame/Cli/Command/CommandLoader.php
@@ -8,13 +8,13 @@ class CommandLoader
     protected array $commands = [];
 
     private const PATHS = [
-        'commands' => __DIR__ . '/../../commands', // TODO: Make this configurable in config
+        'commands' => __DIR__ . '/../../../../app/commands', // Points to app/commands directory  // TODO: Test user defined commands
         'builtins' => __DIR__ . '/BuiltIns',
     ];
 
     private const NAMESPACES = [
-        'commands' => 'commands', // TODO: Make this configurable in config
-        'builtins' => 'Frame\Cli\Command\BuiltIns',
+        'commands' => 'App\\Commands', // TODO: Test user defined commands
+        'builtins' => 'Frame\Cli\Command\BuiltIns', // Full namespace path
     ];
 
     public function __construct()
