+++ src/Frame/Controller/BaseController.php
@@ -9,9 +9,10 @@ abstract class BaseController
     }
 
     // TODO: Make config.php or smth to store the template directory among other things
+    // TODO: Make a functions.php file in the framework for common functions
     public static function view(string $view, array $data = [], string $templateDir = 'views'): void
     {
         extract($data);
-        include __DIR__ . "/../../{$templateDir}/{$view}.php";
+        include __DIR__ . "/../../../app/{$templateDir}/{$view}.php";
     }
 }
