--- src/migrations.sql
@@ -1,46 +1,0 @@
--- Create users table
-CREATE TABLE users
-(
-    id         INT AUTO_INCREMENT PRIMARY KEY,
-    naam       VARCHAR(255) NOT NULL,
-    email      VARCHAR(255) NOT NULL UNIQUE,
-    wachtwoord VARCHAR(255) NOT NULL,
-    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
-);
-
--- Create tasks table
-CREATE TABLE tasks
-(
-    id           INT AUTO_INCREMENT PRIMARY KEY,
-    title        VARCHAR(255) NOT NULL,
-    beschrijving TEXT,
-    status       VARCHAR(50)  NOT NULL,
-    verval_datum INT UNSIGNED NOT NULL,
-    prioriteit   VARCHAR(50)  NOT NULL,
-    created_at   TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at   TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
-);
-
--- Create sub_tasks table
-CREATE TABLE sub_tasks
-(
-    id           INT AUTO_INCREMENT PRIMARY KEY,
-    task_id      INT          NOT NULL,
-    title        VARCHAR(255) NOT NULL,
-    beschrijving TEXT,
-    status       VARCHAR(50)  NOT NULL,
-    created_at   TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    updated_at   TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-    FOREIGN KEY (task_id) REFERENCES tasks (id) ON DELETE CASCADE
-);
-
--- Create user_task pivot table for many-to-many relationship
-CREATE TABLE task_user
-(
-    user_id INT NOT NULL,
-    task_id INT NOT NULL,
-    PRIMARY KEY (user_id, task_id),
-    FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE CASCADE,
-    FOREIGN KEY (task_id) REFERENCES tasks (id) ON DELETE CASCADE
-);
